requirements.txt
EbookLib==0.17.1
Pillow==9.3.0
playwright==1.28.0
pyquery==1.4.3
requests==2.28.1
